+++ src/tools/rust-analyzer/lib/smol_str/.gitignore
@@ -1,3 +1,4 @@
 /target
 /ci
+/.vscode
 Cargo.lock
