+++ src/tools/rust-analyzer/lib/smol_str/Cargo.toml
@@ -5,21 +5,22 @@ description = "small-string optimized string type with O(1) clone"
 license = "MIT OR Apache-2.0"
 repository = "https://github.com/rust-analyzer/smol_str"
 authors = ["Aleksey Kladov <aleksey.kladov@gmail.com>"]
-edition = "2018"
+edition = "2021"
 
 [package.metadata.docs.rs]
 rustdoc-args = ["--cfg", "docsrs"]
 all-features = true
 
 [dependencies]
-serde = { version = "1.0.136", optional = true, default-features = false }
-arbitrary = { version = "1.1.0", optional = true }
+serde = { version = "1.0", optional = true, default-features = false }
+arbitrary = { version = "1.3", optional = true }
 
 [dev-dependencies]
-proptest = "1.0.0"
-serde_json = "1.0.79"
-serde = { version = "1.0.136", features = ["derive"] }
+proptest = "1.5"
+serde_json = "1.0"
+serde = { version = "1.0", features = ["derive"] }
 
 [features]
 default = ["std"]
 std = ["serde?/std"]
+serde = ["dep:serde"]
